cdk-java: reorganize the DestinationAccptanceTest to split out the actual tests from all the util methods
