+++ airbyte-cdk/java/airbyte-cdk/core/src/main/kotlin/io/airbyte/cdk/integrations/destination/async/state/GlobalAsyncStateManager.kt
@@ -108,6 +108,7 @@ class GlobalAsyncStateManager(private val memoryManager: GlobalMemoryManager) {
             convertToGlobalIfNeeded(message)
             preState = false
         }
+        logger.info {"SGX message=$message"}
         // stateType should not change after a conversion.
         Preconditions.checkArgument(stateType == extractStateType(message))
